requirements.txt
qdrant-client
sentence-transformers
python-dotenv
torch